purge TrackedModelWrapper
